separate query_radius
